+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views.cljs
@@ -38,9 +38,9 @@
   [{:keys [step-id] :as step-state}]
   (let [cred-id                (subscribe [::subs/selected-credential-id])
         credential-loading?    (subscribe [::creds-subs/credential-check-loading? @cred-id])
-        credential-invalid?    (subscribe [::creds-subs/credential-check-status-invalid? @cred-id])
+        credential-valid?    (subscribe [::creds-subs/credential-check-status-valid? @cred-id])
         check-status           (creds-utils/credential-check-status
-                                 @credential-loading? @credential-invalid?)
+                                 @credential-loading? (not @credential-valid?))
         credentials-completed? (subscribe [::subs/credentials-completed?])]
     (dispatch [::events/set-launch-status step-id check-status])
     (or (when (and @credentials-completed? (not= :ok check-status))